Rename old compaction creation test
