+++ java/system-test/system-test-suite/src/test/java/sleeper/systemtest/suite/CompactionCreationThroughputST.java
@@ -20,73 +20,58 @@ import org.junit.jupiter.api.Test;
 
 import sleeper.compaction.core.job.creation.strategy.impl.BasicCompactionStrategy;
 import sleeper.core.partition.PartitionTree;
-import sleeper.core.partition.PartitionsBuilderSplitsFirst;
-import sleeper.core.schema.Schema;
-import sleeper.core.schema.type.IntType;
-import sleeper.core.statestore.FileReference;
-import sleeper.core.statestore.FileReferenceFactory;
-import sleeper.core.statestore.transactionlog.transaction.impl.AddFilesTransaction;
 import sleeper.core.util.PollWithRetries;
-import sleeper.core.util.SplitIntoBatches;
 import sleeper.systemtest.dsl.SleeperSystemTest;
+import sleeper.systemtest.dsl.compaction.FoundCompactionJobs;
+import sleeper.systemtest.suite.testutil.Slow;
 import sleeper.systemtest.suite.testutil.SystemTest;
 
 import java.time.Duration;
-import java.util.List;
 import java.util.Map;
-import java.util.stream.IntStream;
-import java.util.stream.Stream;
+import java.util.stream.LongStream;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static sleeper.core.properties.table.TableProperty.COMPACTION_FILES_BATCH_SIZE;
 import static sleeper.core.properties.table.TableProperty.COMPACTION_STRATEGY_CLASS;
 import static sleeper.core.properties.table.TableProperty.TABLE_ONLINE;
-import static sleeper.core.schema.SchemaTestHelper.schemaWithKey;
-import static sleeper.core.statestore.testutils.StateStoreUpdatesWrapper.update;
-import static sleeper.systemtest.suite.fixtures.SystemTestInstance.MAIN;
+import static sleeper.systemtest.dsl.testutil.SystemTestPartitionsTestHelper.createPartitionTreeWithRecordsPerPartitionAndTotal;
+import static sleeper.systemtest.dsl.util.SystemTestSchema.DEFAULT_SCHEMA;
+import static sleeper.systemtest.suite.fixtures.SystemTestInstance.COMPACTION_CREATION;
 
 @SystemTest
+@Slow // Slow because it deploys its own instance so it can drain the whole compaction jobs queue
 public class CompactionCreationThroughputST {
-
-    int numCompactions = 200_000;
-
-    Schema schema = schemaWithKey("key", new IntType());
-    List<String> partitionIds = IntStream.rangeClosed(1, numCompactions)
-            .mapToObj(i -> "partition-" + i).toList();
-    List<Object> splits = IntStream.range(1, numCompactions).mapToObj(i -> (Object) i).toList();
-    PartitionTree partitions = PartitionsBuilderSplitsFirst
-            .leavesWithSplits(schema, partitionIds, splits)
-            .anyTreeJoiningAllLeaves().buildTree();
-    FileReferenceFactory fileFactory = FileReferenceFactory.from(partitions);
 
     @BeforeEach
     void setUp(SleeperSystemTest sleeper) {
-        sleeper.connectToInstanceNoTables(MAIN);
+        sleeper.connectToInstanceNoTables(COMPACTION_CREATION);
     }
 
     @Test
-    void shouldFakeCompactionCreation(SleeperSystemTest sleeper) {
+    void shouldCreateLargeQuantitiesOfCompactionJobsAtOnce(SleeperSystemTest sleeper) throws Exception {
         // Given
-        sleeper.tables().createWithProperties("compaction", schema, Map.of(
+        sleeper.tables().createWithProperties("test", DEFAULT_SCHEMA, Map.of(
                 TABLE_ONLINE, "false",
                 COMPACTION_STRATEGY_CLASS, BasicCompactionStrategy.class.getName(),
                 COMPACTION_FILES_BATCH_SIZE, "2"));
+        PartitionTree partitions = createPartitionTreeWithRecordsPerPartitionAndTotal(10, 655360, sleeper);
         sleeper.partitioning().setPartitions(partitions);
-        sleeper.stateStore().fakeCommits().setupStateStore(store -> {
-            Stream<FileReference> files = IntStream.rangeClosed(1, numCompactions).mapToObj(i -> i)
-                    .flatMap(i -> Stream.of(
-                            fileFactory.partitionFile("partition-" + i, "compaction-" + i + "file-1.parquet", 100),
-                            fileFactory.partitionFile("partition-" + i, "compaction-" + i + "file-2.parquet", 200)));
-            SplitIntoBatches.streamBatchesOf(10000, files)
-                    .map(AddFilesTransaction::fromReferences)
-                    .forEach(update(store)::addTransaction);
-        });
+        sleeper.sourceFiles().inDataBucket().writeSketches()
+                .createWithNumberedRecords("file1.parquet", LongStream.range(0, 655360))
+                .createWithNumberedRecords("file2.parquet", LongStream.range(0, 655360));
+        sleeper.ingest().toStateStore()
+                .addFileOnEveryPartition("file1.parquet", 655360)
+                .addFileOnEveryPartition("file2.parquet", 655360);
 
         // When
-        sleeper.compaction().createJobs(numCompactions,
-                PollWithRetries.intervalAndPollingTimeout(Duration.ofSeconds(1), Duration.ofMinutes(1)));
+        FoundCompactionJobs jobs = sleeper.compaction()
+                .putTableOnlineWaitForJobCreation(65536,
+                        PollWithRetries.intervalAndPollingTimeout(Duration.ofSeconds(10), Duration.ofMinutes(5)))
+                .drainJobsQueueForWholeInstance();
 
         // Then
-        // TODO drain compaction jobs queue
+        assertThat(jobs.checkFullCompactionWithPartitionsAndInputFiles(
+                partitions, "file1.parquet", "file2.parquet"))
+                .isEqualTo(FoundCompactionJobs.expectedFullCompactionCheckWithJobs(65536));
     }
-
 }
